Do not convert \s5 into a paragraph node
When converting from USFM to USX, the \s5 tag cannot be brought over as a paragraph node. Since paragraphs cannot be nested, this causes the current paragraph to close prematurely and the remainder of the verses that are supposed to be in the paragraph are orphaned, not in any paragraph.  This is not valid USX,and it was interferring with the script that converts from USFM to translationStudio.